Rename controller generator file
